requirements.txt
# Pin with: pip install -r requirements.txt
google-generativeai>=0.8.6
pandas>=2.3.0
plotly>=6.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
